# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EZ Remote Agent",
    "description": "Installs, signs in and runs coding agents.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List agents",
        "description": "Returns every agent with its install and sign-in state.",
        "operationId": "listAgents",
        "responses": {
          "200": {
            "description": "Agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentStatus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/claude/settings": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Claude Code settings",
        "operationId": "getClaudeSettings",
        "responses": {
          "200": {
            "description": "Settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeSettingsBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "Change Claude Code settings",
        "description": "Saves the settings, never downgrading the installed version and restarting Remote Control when its settings change.",
        "operationId": "updateClaudeSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaudeSettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeSettingsBody"
                }
              }
            }
          },
          "400": {
            "description": "A setting Claude cannot take",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/install": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Install or update an agent",
        "description": "Installs the newest release on the agent's release channel when it is newer than the installed one, and marks the agent configured.",
        "operationId": "installAgent",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The download or checksum failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/login": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Start signing in",
        "description": "Starts the agent's sign-in, replacing any unfinished one, and returns the link to open.",
        "operationId": "startAgentLogin",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sign-in started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPrompt"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The agent is not installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The agent did not show a sign-in link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/login/code": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Finish Claude sign-in",
        "description": "Passes the code Claude shows after website sign-in to the waiting sign-in.",
        "operationId": "submitAgentLoginCode",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent, only claude takes a code",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Signed in"
          },
          "400": {
            "description": "The agent does not take a code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "No sign-in is in progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The sign-in did not finish",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent}/logout": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Sign the agent out",
        "description": "Signs the agent's CLI out of its account.",
        "operationId": "logOutAgent",
        "parameters": [
          {
            "name": "agent",
            "in": "path",
            "description": "The agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Agent"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The agent is not installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The CLI could not sign out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Stream what changes",
        "description": "Sends `connected`, then each part that changes, until the manager shuts down.",
        "operationId": "streamEvents",
        "responses": {
          "200": {
            "description": "Server-sent events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerEvent"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/folders": {
      "get": {
        "tags": [
          "folders"
        ],
        "summary": "List the folders in /projects",
        "operationId": "listFolders",
        "responses": {
          "200": {
            "description": "Folders sorted by name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderStatus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/folders/{name}/remote-control": {
      "put": {
        "tags": [
          "folders"
        ],
        "summary": "Choose whether the Claude app lists a folder",
        "description": "Starts or stops the folder's Remote Control server.",
        "operationId": "chooseToServeFolder",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The folder's name in /projects",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Saved"
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Get the GitHub sign-in and commit identity",
        "operationId": "getGitStatus",
        "responses": {
          "200": {
            "description": "Git status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/github/login": {
      "post": {
        "tags": [
          "git"
        ],
        "summary": "Start GitHub sign-in",
        "description": "Starts a device sign-in and returns the page to open and the code to enter there.",
        "operationId": "startGitHubLogin",
        "responses": {
          "200": {
            "description": "Sign-in started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPrompt"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The sign-in comes from an environment variable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "gh did not start a sign-in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/github/logout": {
      "post": {
        "tags": [
          "git"
        ],
        "summary": "Sign git out of GitHub",
        "operationId": "logOutOfGitHub",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The sign-in comes from an environment variable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "gh could not sign out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git/identity": {
      "put": {
        "tags": [
          "git"
        ],
        "summary": "Set the commit name and email",
        "description": "An empty or missing value removes it.",
        "operationId": "updateCommitIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitIdentity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitIdentity"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "git could not save it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/login": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Sign in",
        "description": "Checks the manager password and signs this browser in.",
        "operationId": "logIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Signed in"
          },
          "401": {
            "description": "Wrong password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "No password is set yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/remote-control": {
      "get": {
        "tags": [
          "remote-control"
        ],
        "summary": "Get every Remote Control server's state",
        "operationId": "getRemoteControl",
        "responses": {
          "200": {
            "description": "The servers for /projects and its folders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteControlOverview"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in to the manager",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/setup": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Set the password",
        "description": "Sets the manager password when none is set yet, and signs this browser in.",
        "operationId": "setUpPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password set and signed in"
          },
          "400": {
            "description": "Empty password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "A password is already set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "string",
        "description": "A command-line coding agent the manager can install and run.",
        "enum": [
          "claude",
          "codex"
        ]
      },
      "AgentStatus": {
        "type": "object",
        "description": "One agent's install and sign-in state.",
        "required": [
          "agent",
          "configured",
          "logged_in"
        ],
        "properties": {
          "account": {
            "type": [
              "string",
              "null"
            ],
            "description": "The signed-in account as the CLI reports it, absent when unknown."
          },
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "available_update": {
            "type": [
              "string",
              "null"
            ],
            "description": "A newer release than the installed version, absent when none is known."
          },
          "config_disk_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Bytes the agent keeps on /config, absent when they cannot be measured.",
            "minimum": 0
          },
          "configured": {
            "type": "boolean",
            "description": "Installed through the manager at least once."
          },
          "install_progress": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DownloadProgress",
                "description": "Present while an install or update is downloading."
              },
              {
                "type": "null"
              }
            ]
          },
          "installed_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Absent when the agent is not installed."
          },
          "logged_in": {
            "type": "boolean"
          },
          "login_prompt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LoginPrompt",
                "description": "Present while a sign-in is waiting for the person signing in."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ClaudeSettingsBody": {
        "type": "object",
        "description": "How the manager installs, updates and serves Claude Code.",
        "required": [
          "release_channel",
          "remote_control"
        ],
        "properties": {
          "release_channel": {
            "$ref": "#/components/schemas/ReleaseChannel"
          },
          "remote_control": {
            "$ref": "#/components/schemas/RemoteControlSettings"
          }
        }
      },
      "CodeBody": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "CommitIdentity": {
        "type": "object",
        "description": "The name and email git writes into commits.",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DownloadProgress": {
        "type": "object",
        "description": "How far a download has got.",
        "required": [
          "received_bytes"
        ],
        "properties": {
          "received_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes received so far.",
            "minimum": 0
          },
          "total_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Size of the download in bytes, absent when the server does not say.",
            "minimum": 0
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "Why a request failed.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "FolderStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Folder"
          },
          {
            "type": "object",
            "required": [
              "serve"
            ],
            "properties": {
              "serve": {
                "type": "boolean",
                "description": "Chosen to be served to the Claude app, by its own choice or by default."
              }
            }
          }
        ],
        "description": "A folder with what the Claude app sees of it."
      },
...
      "GitStatus": {
        "type": "object",
        "description": "The GitHub sign-in and commit identity git uses.",
        "required": [
          "github",
          "identity"
        ],
        "properties": {
          "github": {
            "$ref": "#/components/schemas/GitHubStatus"
          },
          "identity": {
            "$ref": "#/components/schemas/CommitIdentity"
          }
        }
      },
      "LoginPrompt": {
        "type": "object",
        "description": "What the person signing in needs: a page to open, and sometimes a code to enter there.",
        "required": [
          "url"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "The one-time code to enter on that page, when the sign-in uses one."
          },
          "url": {
            "type": "string",
            "description": "The sign-in page to open."
          }
        }
      },
      "ManagerEvent": {
        "oneOf": [
          {
            "type": "object",
            "description": "Sent first on every connection, with the current revision.",
            "required": [
              "revision",
              "event"
            ],
            "properties": {
              "event": {
                "type": "string",
                "enum": [
                  "connected"
                ]
              },
              "revision": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "description": "This part changed.",
            "required": [
              "topic",
              "revision",
              "event"
            ],
            "properties": {
              "event": {
                "type": "string",
                "enum": [
                  "changed"
                ]
              },
              "revision": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "topic": {
                "$ref": "#/components/schemas/Topic"
              }
            }
          }
        ],
        "description": "One message on the event stream. Every change raises the revision by one."
      },
      "PasswordBody": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "RemoteControlOverview": {
        "type": "object",
        "description": "Every Remote Control server, and where the Claude app lists them.",
        "required": [
          "projects",
          "folders"
        ],
        "properties": {
          "device": {
            "type": [
              "string",
              "null"
            ],
            "description": "The device the Claude app lists the servers under, the container's hostname."
          },
          "folders": {
            "type": "object",
            "description": "The servers for folders in /projects, by folder name.",
            "additionalProperties": {
              "$ref": "#/components/schemas/RemoteControlStatus"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "projects": {
            "$ref": "#/components/schemas/RemoteControlStatus",
            "description": "The server for /projects, waiting when it has not started yet."
          }
        }
      },
...
      "ServeBody": {
        "type": "object",
        "description": "Whether the Claude app lists a folder.",
        "required": [
          "serve"
        ],
        "properties": {
          "serve": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}